chore(wmg): move XAxisChart
