+++ frontend/src/views/WheresMyGeneV2/components/HeatMap/index.tsx
@@ -51,13 +51,13 @@ import {
   XAxisMask,
   YAxisWrapper,
 } from "src/views/WheresMyGene/components/HeatMap/style";
-import { CellCountLabel } from "src/views/WheresMyGene/components/HeatMap/components/XAxisChart/style";
+import { CellCountLabel } from "src/views/WheresMyGeneV2/components/HeatMap/components/XAxisChart/style";
 import {
   HEATMAP_CONTAINER_ID,
   MARGIN_BETWEEN_HEATMAPS,
 } from "src/views/WheresMyGeneV2/common/constants";
 import Loader from "src/views/WheresMyGene/components/Loader";
-import XAxisChart from "src/views/WheresMyGene/components/HeatMap/components/XAxisChart";
+import XAxisChart from "src/views/WheresMyGeneV2/components/HeatMap/components/XAxisChart";
 import Chart from "src/views/WheresMyGeneV2/components/HeatMap/components/Chart";
 import { hyphenize } from "src/views/WheresMyGene/components/HeatMap/utils";
 import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "../GeneSearchBar/components/SaveExport";
